Updated project structure (#4)
* Updated project structure

* Project is running now

* Moved package.swift

* Deleted tests

* Added comments for some functions

* Added more comments

* Moar comments

* More rejigging

* Renamed a couple of things

* Moar documentation

* Updated things
