+++ Examples/DemoApp/ColorGenerator.xcodeproj/project.xcworkspace/contents.xcworkspacedata
