+++ Examples/DemoApp/ColorGenerator.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
