+++ Examples/DemoApp/ColorGenerator/Assets.xcassets/AccentColor.colorset/Contents.json
