+++ Examples/DemoApp/ColorGenerator/Assets.xcassets/AppIcon.appiconset/Contents.json
