+++ Examples/DemoApp/ColorGenerator/Assets.xcassets/Contents.json
