+++ Examples/DemoApp/ColorGenerator/ColorGenerator.entitlements
