+++ Examples/DemoApp/ColorGenerator/ColorGeneratorApp.swift
