+++ Examples/DemoApp/ColorGenerator/Features/ColorTokensRamp/ColorTokensGeneratorViewModel.swift
