+++ Examples/DemoApp/ColorGenerator/Features/Hues/HuesView.swift
