+++ Examples/DemoApp/ColorGenerator/Features/Hues/HuesViewModel.swift
