+++ Examples/DemoApp/ColorGenerator/Features/Navigation/GlobalRouter.swift
