+++ Examples/DemoApp/ColorGenerator/Features/Navigation/MainView.swift
