+++ Examples/DemoApp/ColorGenerator/Preview Content/Preview Assets.xcassets/Contents.json
