+++ Examples/DemoApp/ColorGenerator/Shared/Color+Ext.swift
