+++ Examples/DemoApp/ColorGenerator/Shared/Double+Ext.swift
