+++ Examples/DemoApp/ColorGenerator/Shared/LCHColor+Ext.swift
