+++ Examples/DemoApp/ColorGenerator/Shared/View+Ext.swift
