+++ Sources/ColorTokensKit/Shared/Constants.swift
@@ -1,0 +1,18 @@
+//
+//  Constants.swift
+//
+//
+//  Created by Siddhant Mehta on 2024-06-10.
+//
+
+import Foundation
+
+enum Constants {
+    static let RAD_TO_DEG = 180 / Double.pi
+    static let LAB_E: CGFloat = 0.008856
+    static let LAB_16_116: CGFloat = 0.1379310
+    static let LAB_K_116: CGFloat = 7.787036
+    static let LAB_X: CGFloat = 0.95047
+    static let LAB_Y: CGFloat = 1
+    static let LAB_Z: CGFloat = 1.08883
+}
